Fix variable type
Line 2214:
Comparison of integers of different signs: 's32' (aka 'int') and 'unsigned long'clang(-Wsign-compare)

+++ src/studio/studio.c
@@ -2211,7 +2211,7 @@ static void processMouseStates(Studio* studio)
     tic->ram->vram.vars.cursor.sprite = tic_cursor_arrow;
     tic->ram->vram.vars.cursor.system = true;
 
-    for(s32 i = 0; i < COUNT_OF(studio->mouse.state); i++)
+    for(size_t i = 0; i < COUNT_OF(studio->mouse.state); i++)
     {
         MouseState* state = &studio->mouse.state[i];
